Slider郡をユーザーコントロールへ移動
