+++ Setting/voicevoxPropertySlider.xaml
@@ -1,0 +1,43 @@
+﻿<UserControl x:Class="Setting.VoicevoxPropertySlider"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Setting"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="500">
+	<UserControl.Resources>
+		<local:DoubleStringConverter x:Key="DoubleConverter1"/>
+		<local:DoubleStringConverter2 x:Key="DoubleConverter2"/>
+	</UserControl.Resources>
+    <StackPanel>
+		<GroupBox Header="音量・話速">
+			<StackPanel>
+				<RadioButton Margin="3" IsChecked="True" Content="SAPIから取得する"/>
+				<RadioButton x:Name="thisRadio" Margin="3" Content="この画面の値を使用する"/>
+			</StackPanel>
+		</GroupBox>
+		<StackPanel IsEnabled="{Binding IsChecked, ElementName=thisRadio}">
+			<StackPanel Orientation="Horizontal">
+				<TextBlock Text="音量" Margin="3" Foreground="Gray" IsEnabledChanged="TextBlock_IsEnabledChanged"/>
+				<TextBlock Text="{Binding Value, ElementName=volumeSlider}" Margin="3" Foreground="Gray" IsEnabledChanged="TextBlock_IsEnabledChanged"/>
+			</StackPanel>
+			<Slider x:Name="volumeSlider" Minimum="0" Maximum="100" SmallChange="1" LargeChange="5" TickFrequency="1" TickPlacement="None" IsSnapToTickEnabled="True" Value="100"/>
+			<StackPanel Orientation="Horizontal">
+				<TextBlock Text="話速" Margin="3" Foreground="Gray" IsEnabledChanged="TextBlock_IsEnabledChanged"/>
+				<TextBlock Text="{Binding Value, ElementName=speedSlider, Converter={StaticResource DoubleConverter1}}" Margin="3" Foreground="Gray" IsEnabledChanged="TextBlock_IsEnabledChanged"/>
+			</StackPanel>
+			<Slider x:Name="speedSlider" Minimum="0.5" Maximum="2.0" SmallChange="0.1" LargeChange="0.1" TickFrequency="0.1" TickPlacement="None" IsSnapToTickEnabled="True" Value="1"/>
+		</StackPanel>
+		<StackPanel Orientation="Horizontal">
+			<TextBlock Text="音高" Margin="3"/>
+			<TextBlock Text="{Binding Value, ElementName=pitchSlider, Converter={StaticResource DoubleConverter2}}" Margin="3"/>
+		</StackPanel>
+		<Slider x:Name="pitchSlider" Maximum="0.15" Minimum="-0.15" SmallChange="0.01" LargeChange="0.01" TickFrequency="0.01" TickPlacement="None" IsSnapToTickEnabled="True"/>
+		<StackPanel Orientation="Horizontal">
+			<TextBlock Text="抑揚" Margin="3"/>
+			<TextBlock Text="{Binding Value, ElementName=intonationSlider, Converter={StaticResource DoubleConverter1}}" Margin="3"/>
+		</StackPanel>
+		<Slider x:Name="intonationSlider" Maximum="2.0" Minimum="0" SmallChange="0.1" LargeChange="0.1" TickFrequency="0.1" TickPlacement="None" IsSnapToTickEnabled="True" Value="1"/>
+	</StackPanel>
+</UserControl>
